Separating the tests.
